CMOS-49: Provide native build pipeline (#19)
Now invokes build and test of the OSS container using native (and K8S but that ignores errors)

+++ testing/microlith-test/kubernetes/standalone.bats
@@ -34,9 +34,13 @@ setup() {
 }
 
 teardown() {
-    run helm uninstall --namespace "${TEST_NAMESPACE}" couchbase
-    run kubectl delete --force --grace-period=0 --now=true --wait=true -n "$TEST_NAMESPACE" -f "$TEST_KUBERNETES_RESOURCES_ROOT/default-microlith.yaml"
-    run kubectl delete namespace "$TEST_NAMESPACE"
+    if [ "$SKIP_TEARDOWN" == "true" ]; then
+        skip "Skipping teardown"
+    elif [ "$TEST_NATIVE" != "true" ]; then
+        run helm uninstall --namespace "${TEST_NAMESPACE}" couchbase
+        run kubectl delete --force --grace-period=0 --now=true --wait=true -n "$TEST_NAMESPACE" -f "$TEST_KUBERNETES_RESOURCES_ROOT/default-microlith.yaml"
+        run kubectl delete namespace "$TEST_NAMESPACE"
+    fi
 }
 
 TEST_NAMESPACE=${TEST_NAMESPACE:-test}
